# api-gateway/static/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library Management System API",
        "description": "A complete API collection for the Library Management System with User, Book, and Category microservices",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://{{api_url}}"
        },
        {
            "url": "http://{{user_service_url}}"
        },
        {
            "url": "http://{{book_service_url}}"
        },
        {
            "url": "http://{{category_service_url}}"
        }
    ],
...
    "paths": {
...
        "/api/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create User (Admin)",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "type": "string",
                                "example": "\"{\\n  \\\"email\\\": \\\"admin@example.com\\\",\\n  \\\"username\\\": \\\"admin\\\",\\n  \\\"password\\\": \\\"AdminPass123\\\",\\n  \\\"first_name\\\": \\\"Admin\\\",\\n  \\\"last_name\\\": \\\"User\\\",\\n  \\\"role\\\": \\\"admin\\\",\\n  \\\"phone\\\": \\\"123-456-7890\\\",\\n  \\\"address\\\": \\\"123 Admin St\\\"\\n}\""
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {{ACCESS_TOKEN}}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List Users (Admin)",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {{ACCESS_TOKEN}}"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List Categories",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Create Category (Admin)",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "type": "string",
                                "example": "\"{\\n  \\\"name\\\": \\\"New Category\\\",\\n  \\\"description\\\": \\\"Description of the new category\\\",\\n  \\\"parent_id\\\": null\\n}\""
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {{ACCESS_TOKEN}}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "List Books",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create Book (Admin)",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "type": "string",
                                "example": "\"{\\n  \\\"title\\\": \\\"New Book Title\\\",\\n  \\\"author\\\": \\\"Author Name\\\",\\n  \\\"isbn\\\": \\\"9781234567890\\\",\\n  \\\"published_year\\\": 2022,\\n  \\\"publisher\\\": \\\"Publisher Name\\\",\\n  \\\"description\\\": \\\"Book description\\\",\\n  \\\"category_ids\\\": [\\\"50c3ef9e-d1aa-4e88-aa75-7d92c9d11111\\\"],\\n  \\\"language\\\": \\\"English\\\",\\n  \\\"page_count\\\": 250,\\n  \\\"cover_image\\\": \\\"https://example.com/cover.jpg\\\",\\n  \\\"quantity\\\": 5\\n}\""
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Content-Type",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "application/json"
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "Bearer {{ACCESS_TOKEN}}"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}